新的管理类、广播类完成。
