+++ src/main/java/ice/games/snake/SnakeGame.java
@@ -4,20 +4,17 @@ import java.io.IOException;
 
 import org.atmosphere.cpr.AtmosphereResource;
 import org.atmosphere.cpr.AtmosphereResourceFactory;
-import org.atmosphere.cpr.BroadcasterFactory;
 
 public class SnakeGame {
 
-	private final SnakeBroadcaster snakeBroadcaster;
 	private OnopenProccesor onopenProccesor;
 	private OncloseProccesor oncloseProccesor;
 	private SnakeManager snakeManager;
 
 	public SnakeGame() {
 		this.snakeManager = new SnakeManager();
-		snakeBroadcaster = new SnakeBroadcaster(BroadcasterFactory.getDefault().lookup("/snake", true), snakeManager);
-		this.onopenProccesor = new OnopenProccesor(snakeBroadcaster, snakeManager);
-		this.oncloseProccesor = new OncloseProccesor(snakeBroadcaster);
+		this.onopenProccesor = new OnopenProccesor(snakeManager);
+		this.oncloseProccesor = new OncloseProccesor(snakeManager);
 	}
 
 	public void onOpen(AtmosphereResource resource) throws IOException {
