+++ src/main/java/ice/games/snake/SnakeManagedService.java
@@ -6,11 +6,9 @@ import java.util.concurrent.ConcurrentLinkedQueue;
 import org.atmosphere.config.service.Get;
 import org.atmosphere.config.service.ManagedService;
 import org.atmosphere.config.service.Post;
-import org.atmosphere.cpr.AtmosphereRequest;
 import org.atmosphere.cpr.AtmosphereResource;
 import org.atmosphere.cpr.AtmosphereResourceEvent;
 import org.atmosphere.cpr.AtmosphereResourceEventListenerAdapter;
-import org.atmosphere.cpr.HeaderConfig;
 
 @ManagedService(path = "/snake")
 public class SnakeManagedService extends SnakeGame {
@@ -35,8 +33,9 @@ public class SnakeManagedService extends SnakeGame {
 
 			@Override
 			public void onDisconnect(AtmosphereResourceEvent event) {
-				AtmosphereRequest request = event.getResource().getRequest();
-				String s = request.getHeader(HeaderConfig.X_ATMOSPHERE_TRANSPORT);
+				// AtmosphereRequest request = event.getResource().getRequest();
+				// String s =
+				// request.getHeader(HeaderConfig.X_ATMOSPHERE_TRANSPORT);
 				// if (s != null && s.equalsIgnoreCase(HeaderConfig.DISCONNECT))
 				// {
 				SnakeManagedService.super.onClose(resource);
