+++ src/main/java/ice/games/snake/SnakeManager.java
@@ -2,9 +2,12 @@ package ice.games.snake;
 
 import java.util.Collection;
 import java.util.Collections;
+import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.concurrent.ConcurrentHashMap;
 
+import org.atmosphere.cpr.AtmosphereResource;
+import org.atmosphere.cpr.BroadcasterFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -15,6 +18,14 @@ public class SnakeManager {
 	private final ConcurrentHashMap<Integer, Snake> playingSnakes = new ConcurrentHashMap<Integer, Snake>();
 
 	private final LinkedList<Snake> waitqueue = new LinkedList<Snake>();
+
+	private static final int MAX_ALIVE_SNAKE = 2;
+
+	private NewSnakeBroadcaster snakeBroadcaster;
+
+	public SnakeManager() {
+		this.snakeBroadcaster = new NewSnakeBroadcaster(BroadcasterFactory.getDefault().lookup("/snake", true), this);
+	}
 
 	Collection<Snake> getPlayingSnakes() {
 		return Collections.unmodifiableCollection(playingSnakes.values());
@@ -49,5 +60,55 @@ public class SnakeManager {
 
 	boolean removeWaitSnake(Snake snake) { // TODO 当等待的人多了以后，这里可能会出问题
 		return waitqueue.remove(snake);
+	}
+
+	public void removeOffLineSnake(AtmosphereResource resource) {
+		snakeBroadcaster.removeAtmosphereResource(resource);
+		Snake snake = (Snake) resource.session().getAttribute("snake");
+		removePlayingSnake(snake.getId());
+		removeWaitSnake(snake);
+
+		Integer snakeId = (Integer) resource.session().getAttribute("id");
+		snakeBroadcaster.broadcast(String.format("{'type': 'leave', 'id': %d}", snakeId));
+	}
+
+	synchronized void addSnake(Snake snake) {
+		waitForPlay(snake);
+		if (playingSnakesNum() >= MAX_ALIVE_SNAKE) {
+			snake.sendMessage(String.format("{'type': 'wait', 'data' : '请稍等,蛇满为患了,您前面还有  %s 条小蛇蛇在焦急等待...'}",
+					waitSnakesNum() - 1)); // 因为一来，就进等待队列，所以里面总会多一个。
+		} else {
+			activeWaitSnake();
+		}
+	}
+
+	synchronized void removeDeadSnake(Snake snake) {
+		int snakeId = snake.getId();
+		removePlayingSnake(snakeId);
+		snakeBroadcaster.broadcast(String.format("{'type': 'dead', 'id': %d}", snakeId));
+		addSnake(snake);
+		activeWaitSnake();
+	}
+
+	private void activeWaitSnake() {
+		if (waitSnakesNum() > 0) {
+			Snake firstWait = getFirstWaitSnake();
+			firstWait.startPlay();
+			addNewPlayingSnake(firstWait);
+			broadcastPlayingSnakeInfo();
+		}
+	}
+
+	private void broadcastPlayingSnakeInfo() {
+		Snake snake;
+		StringBuilder sb = new StringBuilder();
+		for (Iterator<Snake> iterator = getPlayingSnakes().iterator(); iterator.hasNext();) {
+			snake = iterator.next();
+			sb.append(String.format("{id: %d, color: '%s'}", Integer.valueOf(snake.getId()), snake.getHexColor()));
+			if (iterator.hasNext()) {
+				sb.append(',');
+			}
+		}
+		snakeBroadcaster.broadcast(String.format("{'type': 'join','data':[%s]}", sb.toString()));
 	}
 }
